refactor: cleanup styling, types, constants, reorganize exports
