--- apps/studio/src/components/menu-bar/index.ts
@@ -1,5 +1,0 @@
-// Copyright 2026 The Lynx Authors. All rights reserved.
-// Licensed under the Apache License Version 2.0 that can be found in the
-// LICENSE file in the root directory of this source tree.
-
-export { MenuBar } from './menu-bar.tsx';
