--- apps/studio/src/components/mockup-motion/index.ts
@@ -1,9 +1,0 @@
-// Copyright 2026 The Lynx Authors. All rights reserved.
-// Licensed under the Apache License Version 2.0 that can be found in the
-// LICENSE file in the root directory of this source tree.
-
-export {
-  MotionStage,
-  MotionStageContainer,
-  MotionPresentation,
-} from '@dugyu/luna-stage/motion';
