--- apps/studio/src/constants/presentation.ts
@@ -1,10 +1,0 @@
-// Copyright 2026 The Lynx Authors. All rights reserved.
-// Licensed under the Apache License Version 2.0 that can be found in the
-// LICENSE file in the root directory of this source tree.
-
-import type { LunaThemeMode, LunaThemeVariant } from '@dugyu/luna-core';
-
-export const STARTING_MODE: LunaThemeMode = 'dark';
-export const STARTING_VARIANT: LunaThemeVariant = 'lunaris';
-
-export const RECORD_MODE = false;
