--- apps/studio/src/types/stage.ts
@@ -1,15 +1,0 @@
-// Copyright 2026 The Lynx Authors. All rights reserved.
-// Licensed under the Apache License Version 2.0 that can be found in the
-// LICENSE file in the root directory of this source tree.
-
-import type { LunaThemeVariant } from '@dugyu/luna-core';
-
-export type MoonriseEvent =
-  | { field: 'luna-variant'; value: LunaThemeVariant }
-  | {
-    field: 'light-mode';
-    value: boolean;
-  }
-  | { field: 'autoplay'; value: boolean }
-  | { field: 'trust'; value: boolean }
-  | { field: 'subscribe'; value: boolean };
